+++ app/src/main/AndroidManifest.xml
@@ -17,7 +17,7 @@
         android:theme="@style/Theme.Moveto"
         tools:targetApi="31">
         <activity
-            android:name="com.maserplay.appai.Login_games_m2023_ru_Activity"
+            android:name="com.maserplay.appai.login.LoginActivity"
             android:exported="false"
             android:label="@string/login_login"
             android:parentActivityName="com.maserplay.appai.SettingsActivity"
