+++ app/src/main/java/com/maserplay/appai/login/LoginClass.kt
@@ -1,0 +1,14 @@
+package com.maserplay.appai.login
+
+class LoginClass(_name: String, _password: String) {
+    private val email: String
+    private val password: String
+
+    init {
+        email = _name
+        password = _password
+    }
+
+}
+
+
