+++ app/src/main/java/com/maserplay/appai/login/LoginResponseClass.kt
@@ -1,0 +1,14 @@
+package com.maserplay.appai.login
+
+class LoginResponseClass {
+    private var email: Boolean? = null
+        get() = field
+    private var password: Boolean? = null
+        get() = field
+    private var hasaccept: Boolean? = null
+        get() = field
+    private var  acceptcookkie: String? = null
+        get() = field
+    private var  accept: LoginClass? = null
+        get() = field
+}
