+++ app/src/main/java/com/maserplay/appai/login/LoginVerifyClass.kt
@@ -1,0 +1,6 @@
+package com.maserplay.appai.login
+
+class LoginVerifyClass {
+    private var verify: Boolean? = null
+        get() = field
+}
