+++ app/src/main/java/com/maserplay/appai/login/LoginVerifySendClass.kt
@@ -1,0 +1,14 @@
+package com.maserplay.appai.login
+
+class LoginVerifySendClass(_acceptcookie: String, _code: Int?) {
+    private var acceptcookie: String
+    private val code: Int?
+
+    init {
+        acceptcookie = _acceptcookie
+        code = _code
+    }
+    constructor(_acceptcookie: String) : this(_acceptcookie, _code = null) {
+        this.acceptcookie = _acceptcookie
+    }
+}
