+++ app/src/main/res/layout/activity_login.xml
@@ -4,7 +4,7 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.maserplay.appai.Login_games_m2023_ru_Activity"
+    tools:context="com.maserplay.appai.login.LoginActivity"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
